# SpanBuilder: chain modifyState functions