Some changes decoupled from https://github.com/typelevel/otel4s/pull/963. 

The changes:
- chain functions
- reuse `InstrumentMeta` instance
- add benchmark

## Benchmark

`addParamsAndUse` is the most relevant benchmark. The throughput is roughly the same, and the normalized memory usage is slightly lower. But realistically, there is no difference. 

### Baseline

```
#### baseline with build use

Benchmark                             (backend)  Mode  Cnt      Score      Error   Units
addParams                                  noop  avgt   80      1.177 ±    0.010   ns/op
addParams:gc.alloc.rate.norm               noop  avgt   80     ≈ 10⁻³               B/op
addParams                              oteljava  avgt   80     61.681 ±    0.221   ns/op
addParams:gc.alloc.rate.norm           oteljava  avgt   80    672.001 ±    9.229    B/op
addParams                                   sdk  avgt   80     71.809 ±    0.275   ns/op
addParams:gc.alloc.rate.norm                sdk  avgt   80    712.002 ±    0.002    B/op
addParamsAndUse                            noop  avgt   80   5673.401 ±  170.063   ns/op
addParamsAndUse:gc.alloc.rate.norm         noop  avgt   80   1004.033 ±    4.615    B/op
addParamsAndUse                        oteljava  avgt   80  11622.442 ± 1421.261   ns/op
addParamsAndUse:gc.alloc.rate.norm     oteljava  avgt   80   7659.956 ±    4.587    B/op
addParamsAndUse                             sdk  avgt   80  22796.905 ±  302.181   ns/op
addParamsAndUse:gc.alloc.rate.norm          sdk  avgt   80  46600.915 ±  628.481    B/op
createSpanBuilder                          noop  avgt   80      0.724 ±    0.008   ns/op
createSpanBuilder:gc.alloc.rate.norm       noop  avgt   80     ≈ 10⁻⁴               B/op
createSpanBuilder                      oteljava  avgt   80      3.884 ±    0.013   ns/op
createSpanBuilder:gc.alloc.rate.norm   oteljava  avgt   80     64.000 ±    0.001    B/op
createSpanBuilder                           sdk  avgt   80     11.018 ±    0.138   ns/op
createSpanBuilder:gc.alloc.rate.norm        sdk  avgt   80     72.000 ±    0.001    B/op
```

### This branch

```
Benchmark                             (backend)  Mode  Cnt      Score      Error   Units
addParams                                  noop  avgt   80      1.159 ±    0.022   ns/op
addParams:gc.alloc.rate.norm               noop  avgt   80     ≈ 10⁻³               B/op
addParams                              oteljava  avgt   80      9.614 ±    0.163   ns/op
addParams:gc.alloc.rate.norm           oteljava  avgt   80    112.000 ±    0.001    B/op
addParams                                   sdk  avgt   80      9.706 ±    0.130   ns/op
addParams:gc.alloc.rate.norm                sdk  avgt   80    112.000 ±    0.001    B/op
addParamsAndUse                            noop  avgt   80   5863.874 ±  112.985   ns/op
addParamsAndUse:gc.alloc.rate.norm         noop  avgt   80   1015.977 ±    0.227    B/op
addParamsAndUse                        oteljava  avgt   80  11447.127 ± 1114.902   ns/op
addParamsAndUse:gc.alloc.rate.norm     oteljava  avgt   80   7744.387 ±    0.337    B/op
addParamsAndUse                             sdk  avgt   80  22802.965 ±  282.893   ns/op
addParamsAndUse:gc.alloc.rate.norm          sdk  avgt   80  45765.508 ±  615.943    B/op
createSpanBuilder                          noop  avgt   80      0.750 ±    0.014   ns/op
createSpanBuilder:gc.alloc.rate.norm       noop  avgt   80     ≈ 10⁻⁴               B/op
createSpanBuilder                      oteljava  avgt   80      2.439 ±    0.011   ns/op
createSpanBuilder:gc.alloc.rate.norm   oteljava  avgt   80     40.000 ±    0.001    B/op
createSpanBuilder                           sdk  avgt   80      2.458 ±    0.010   ns/op
createSpanBuilder:gc.alloc.rate.norm        sdk  avgt   80     40.000 ±    0.001    B/op
```